apps/user/tests/factories.py
import factory
from apps.user.models import CustomUser
class UserFactory(factory.django.DjangoModelFactory):
    class Meta:
        model = CustomUser

    username = factory.Faker('user_name')
    email = factory.Faker('email')
