pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["sudip-khanal <sudipkhanal0777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
django-filter = "^24.3"
pypdf2 = "^3.0.1"
faker = "^27.4.0"
requests = "^2.32.3"
celery = "^5.4.0"
django-crontab = "^0.7.1"
redis = "^5.0.8"
drf-yasg = "^1.21.7"
setuptools = "^73.0.1"
django-redis = "^5.4.0"
pytest-django = "^4.8.0"
factory-boy = "^3.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
